refactor(pay): 重构支付模块的类型定义和命名
- 将 PayChannelTyping 重命名为 PayChannel，作为支付渠道的枚举类
- 将 PaymentTyping 重命名为 PaymentProviderType，作为支付提供者类型的枚举类
- 更新相关文件的引用，以适应新的类名
- 从 typing 包移动 ISO4217 和 EncryptAlgorithm 到 enums 包
